refactor(mercury): split channel_bundle/channel along type + lifecycle seams
Pure code movement to bring two oversized files under the 700-line hard
cap (CLAUDE.md "File organization"). No logic, signature, or behavior
changes; re-export discipline keeps every external
`use cimmeria_mercury::...` path identical.

channel_bundle.rs (1018) -> channel_bundle/ directory:
  - idbase.rs       idbase constants + idbase_from_exposed_method_count
                    formula (+ its formula tests)
  - bundle/mod.rs   the ChannelBundle accumulator + finalize path
  - bundle/tests.rs ChannelBundle encoder/fragmentation/ack tests
                    (child of bundle so the byte-exact assertions keep
                    private `body` access)
  - mod.rs          module doc + `pub use` of ChannelBundle and the
                    idbase items

channel/mod.rs (773) -> new sibling submodules under channel/:
  - state.rs        ChannelState / TxEntry / RxEntry bookkeeping records
  - channel_core.rs the Channel struct + state-machine impl (struct and
                    impl kept together so private fields stay visible)
  - tick.rs         TickActions tick-driver punch list + contract doc
  - mod.rs          slimmed to declarations + `pub use` re-exports

Net-new test: ChannelBundle::num_acks coverage (was untested).

Part of #529.
